requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
google-auth==1.21.0
google-auth-oauthlib==0.4.1
gspread==3.6.0
idna==2.10
imgurpython==1.1.7
instaloader==4.5.2
joblib==0.16.0
kiwisolver==1.2.0
matplotlib==3.3.1
nltk==3.5
numpy==1.19.1
oauthlib==3.1.0
pandas==1.1.1
Pillow==7.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tqdm==4.48.2
urllib3==1.25.10
wordcloud==1.8.0
